% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bind_cols}
\alias{bind_cols}
\title{Right-bind columns of matrices from in list.}
\usage{
bind_cols(x)
}
\arguments{
\item{x}{A list of matrices}
}
\description{
Right-bind columns of matrices from in list.
}
