% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blockDiag}
\alias{blockDiag}
\title{Construct block-diagonal matrix from list of matrices}
\usage{
blockDiag(x)
}
\arguments{
\item{x}{A list of matrices}
}
\description{
Constructs as dense block-diagonal matrix from a list (field) of matrices.
}
