% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfvb_lm.R
\name{check.mfvb.x}
\alias{check.mfvb.x}
\title{Check the `x` input to an mfvb regression function}
\usage{
check.mfvb.x(x)
}
\arguments{
\item{x}{The x input to check}
}
\value{
x if all checks pass
}
\description{
Check the `x` input to an mfvb regression function
}
