% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfvb_lm.R
\name{check.mfvb.y}
\alias{check.mfvb.y}
\title{Check the `y` input to an mfvb regression function}
\usage{
check.mfvb.y(x, y)
}
\arguments{
\item{x}{The x input}

\item{y}{The y input to check}
}
\value{
x if all checks pass
}
\description{
Check the `y` input to an mfvb regression function
}
