% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{coef.mfvb}
\alias{coef.mfvb}
\title{Extract model coefficients posterior means}
\usage{
\method{coef}{mfvb}(object)
}
\arguments{
\item{object}{a `mfvb` model}
}
\description{
`coef.mfvb` implements the `coef` generic for objects of class `mfvb`.
}
