% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{confint.mfvb}
\alias{confint.mfvb}
\title{Compute credible intervals for parameters of `mfvb`}
\usage{
\method{confint}{mfvb}(object, level = 0.95, ...)
}
\arguments{
\item{object}{a `mfvb` model}

\item{level}{credible interval level}
}
\description{
Compute credible intervals for parameters of `mfvb`
}
