% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dnorm_mat}
\alias{dnorm_mat}
\title{Evaluate standard normal density for matrix of variates}
\usage{
dnorm_mat(m)
}
\arguments{
\item{m}{A matrix of variates}
}
\description{
Evaluate standard normal density for matrix of variates
}
