% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dot_y_minus_Xb}
\alias{dot_y_minus_Xb}
\title{E[||y - Xb||^2]}
\usage{
dot_y_minus_Xb(yty, Xty, XtX, mu, Sigma)
}
\arguments{
\item{yty}{Statistic y'y}

\item{Xty}{Statistic X'y}

\item{XtX}{Statistic X'X}

\item{mu}{Variational mean mu}

\item{Sigma}{Variational variance Sigma}
}
\description{
Calculate E[||y - Xb||^2] for b ~ MVN(mu, Sigma)
}
