% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ExpectationGaussianSufficientStatistics}
\alias{ExpectationGaussianSufficientStatistics}
\title{Gaussian transform natural parameters to E[T(X)] parameters.}
\usage{
ExpectationGaussianSufficientStatistics(eta)
}
\arguments{
\item{eta}{The natural parameter vector.}
}
\description{
Gaussian transform natural parameters to E[T(X)] parameters.
}
