% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ExpectationInverseGWishartSufficientStatistics}
\alias{ExpectationInverseGWishartSufficientStatistics}
\title{Inverse-G-Wishart transform natural parameters to E[T(X)] parameters.}
\usage{
ExpectationInverseGWishartSufficientStatistics(G, eta)
}
\arguments{
\item{G}{The graph.}

\item{eta}{The natural parameter vector.}
}
\description{
Inverse-G-Wishart transform natural parameters to E[T(X)] parameters.
}
