% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{G_VMP}
\alias{G_VMP}
\title{Calculate}
\usage{
G_VMP(v, Q, r, s)
}
\arguments{
\item{Q}{A d x d matrix}

\item{r}{A d x 1 vector}

\item{s}{A double}

\item{v1}{A d x 1 vector}

\item{v2}{A d*d x 1 vector}
}
\description{
Calculate
}
