% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GaussianEntropy}
\alias{GaussianEntropy}
\title{Calculate entropy for Gaussian density}
\usage{
GaussianEntropy(eta)
}
\arguments{
\item{eta}{Natural parameter}
}
\description{
Calculate entropy for Gaussian density
}
