% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GaussianPriorFragment}
\alias{GaussianPriorFragment}
\title{Gaussian prior fragment update}
\usage{
GaussianPriorFragment(mu, Sigma)
}
\arguments{
\item{mu}{The prior mean}

\item{Sigma}{The prior variance}
}
\description{
Gaussian prior fragment update
}
