% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ig_E_log}
\alias{ig_E_log}
\title{Inverse Gamma E[log(x)]}
\usage{
ig_E_log(a, b)
}
\arguments{
\item{a}{shape}

\item{b}{scale}
}
\description{
Calculate and return the entropy for inverse gamma distribution
with supplied shape and scale.
}
