% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ig_E_lpdf}
\alias{ig_E_lpdf}
\title{E_q[ln p(x)] where x ~ IG(a0, b0) and q(x) = IG(x | a, b)}
\usage{
ig_E_lpdf(a0, b0, a, b)
}
\arguments{
\item{a0}{Inverse gamma prior parameter}

\item{b0}{Inverse gamma prior parameter}

\item{a}{Inverse gamma variational parameter}

\item{b}{Inverse gamma variational parameter}
}
\description{
E_q[ln p(x)] where x ~ IG(a0, b0) and q(x) = IG(x | a, b)
}
