% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ig_entropy}
\alias{ig_entropy}
\title{Inverse Gamma Entropy}
\usage{
ig_entropy(a, b)
}
\arguments{
\item{a}{shape}

\item{b}{scale}
}
\description{
Calculate and return the entropy for inverse gamma distribution
with supplied shape and scale.
}
