% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inv_vech}
\alias{inv_vech}
\title{Calculate inverse of vech for a vector}
\usage{
inv_vech(v)
}
\arguments{
\item{v}{A vector of dimension d(d+1)/2 of lower triangular elements}
}
\description{
Calculate inverse of vech for a vector
}
