% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inv_wishart_E_invX}
\alias{inv_wishart_E_invX}
\title{Inverse Wishart E[X^-1]}
\usage{
inv_wishart_E_invX(nu, S)
}
\arguments{
\item{nu}{Degrees of freedom}

\item{S}{Scale matrix}
}
\description{
Calculate E[X^-1] where X ~ Inverse-Wishart(nu, S).
}
