% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inv_wishart_E_logdet}
\alias{inv_wishart_E_logdet}
\title{Inverse Wishart E[log|X|]}
\usage{
inv_wishart_E_logdet(nu, S)
}
\arguments{
\item{nu}{Degrees of freedom}

\item{S}{Scale matrix}
}
\description{
Calculate E[log|X|] where X ~ Inverse-Wishart(nu, S).
}
