% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InverseGammaPriorFragment}
\alias{InverseGammaPriorFragment}
\title{Inverse-Gamma prior fragment update}
\usage{
InverseGammaPriorFragment(a, b)
}
\arguments{
\item{a}{Shape parameter}

\item{b}{Scale parameter}
}
\description{
Inverse-Gamma prior fragment update
}
