% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InverseGWishartCommonParameters}
\alias{InverseGWishartCommonParameters}
\title{Inverse-G-Wishart Common parameters}
\usage{
InverseGWishartCommonParameters(eta)
}
\arguments{
\item{eta}{The natural parameter vector.}
}
\description{
Inverse-G-Wishart Common parameters
}
