% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InverseGWishartPriorFragment}
\alias{InverseGWishartPriorFragment}
\title{Inverse G-Wishart prior fragment update}
\usage{
InverseGWishartPriorFragment(G, xi, Lambda)
}
\arguments{
\item{G}{The graph matrix}

\item{xi}{The prior shape}

\item{Lambda}{The prior symmetric positive definite matrix}
}
\description{
Inverse G-Wishart prior fragment update
}
