% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{InverseWishartPriorFragment}
\alias{InverseWishartPriorFragment}
\title{Inverse Wishart prior fragment update}
\usage{
InverseWishartPriorFragment(xi, Lambda)
}
\arguments{
\item{xi}{The prior shape}

\item{Lambda}{The prior symmetric positive definite matrix}
}
\description{
Inverse Wishart prior fragment update
}
