% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IteratedInverseGWishartFragment}
\alias{IteratedInverseGWishartFragment}
\title{Iterated Inverse G-Wishart fragment update}
\usage{
IteratedInverseGWishartFragment(G, xi, eta1_in, eta2_in)
}
\arguments{
\item{G}{The graph matrix.}

\item{xi}{The prior shape \eqn{\xi \in \mathbb{R}}.}

\item{eta1_in}{A d + 1 x 1 vector.}

\item{eta2_in}{A d + 1 x 1 vector.}
}
\description{
Iterated Inverse G-Wishart fragment update
}
