% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jaakkola_jordan_n}
\alias{jaakkola_jordan_n}
\title{Perform Jaakkola-Jordan update of variational parameters}
\usage{
jaakkola_jordan_n(X, y, n, eta1, eta2, eta1_p, eta2_p)
}
\arguments{
\item{X}{The design matrix}

\item{y}{The response vector}

\item{n}{The trial vector}

\item{eta1}{The current value of 1st natural parameter}

\item{eta2}{The current value of the 2nd natural parameter}

\item{eta1_p}{The prior value of the 1st natural parameter}

\item{eta2_p}{The prior value of the 2nd natural parameter}
}
\description{
Perform Jaakkola-Jordan update of variational parameters
}
