% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jj_logistic_mixed}
\alias{jj_logistic_mixed}
\title{Variational Bayes for logistic mixed model.}
\usage{
jj_logistic_mixed(X, Z, y, mu_beta, sigma_beta, mu, sigma, A = 1,
  E_inv_sigsq = 1, E_inv_a = 1, tol = 1e-08, maxiter = 100L,
  verbose = FALSE)
}
\arguments{
\item{X}{The design matrix}

\item{Z}{Group design matrix}

\item{y}{The response vector}

\item{mu_beta}{The prior mean for beta}

\item{sigma_beta}{The prior covariance for beta}

\item{mu}{Initial value for mu}

\item{sigma}{Initial value for sigma}

\item{A}{Initial value for A}

\item{E_inv_sigsq}{Initial value for E(1/sigma^2)}

\item{E_inv_a}{Initial value for E(1/a)}

\item{tol}{Tolerance level}

\item{maxiter}{Maximum iterations}

\item{verbose}{Print trace of the lower bound to console. Default is \code{FALSE}.}
}
\value{
A list containing:
\describe{
  \item{converged}{Indicator for algorithm convergence.}
  \item{elbo}{Vector of the ELBO sequence.} 
  \item{mu}{The optimised value of mu.}
  \item{Sigma}{The optimised value of Sigma.}
}
}
\description{
Variational Bayes for logistic mixed model.
}
