% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knowles_minka_wand}
\alias{knowles_minka_wand}
\title{Perform Knowles-Minka-Wand update of variational parameters}
\usage{
knowles_minka_wand(X, y, eta1, eta2, eta1_p, eta2_p, MS_p, MS_s)
}
\arguments{
\item{X}{The design matrix}

\item{y}{The response vector}

\item{eta1}{The current value of 1st natural parameter}

\item{eta2}{The current value of the 2nd natural parameter}

\item{eta1_p}{The prior value of the 1st natural parameter}

\item{eta2_p}{The prior value of the 2nd natural parameter}

\item{MS_p}{The}

\item{MS_s}{The}
}
\description{
Perform Knowles-Minka-Wand update of variational parameters
}
