% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lmvgamma}
\alias{lmvgamma}
\title{Log multivariate Gamma function}
\usage{
lmvgamma(x, p = 1L)
}
\arguments{
\item{x}{The value to evaluate}

\item{p}{The dimension}
}
\description{
Log transformation of the multivariate gamma function.
}
