% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfvb_lm.R
\name{mfvb.control}
\alias{mfvb.control}
\title{Control arguments for mfvb functions}
\usage{
mfvb.control(tol = 1e-08, maxiter = 100, verbose = FALSE)
}
\arguments{
\item{tol}{a positive convergence tolerance for ELBO}

\item{maxiter}{integer giving the maximal number of iterations}

\item{verbose}{logical indicating if output should be printed for each iteration}
}
\value{
A list with named components
}
\description{
The `control` argument
}
