% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvdigamma}
\alias{mvdigamma}
\title{Multivariate digamma function}
\usage{
mvdigamma(x, p = 1L)
}
\arguments{
\item{x}{The value to evaluate}

\item{p}{The dimension}
}
\description{
Derivative of the multivariate Gamma function.
}
