% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvgamma}
\alias{mvgamma}
\title{multivariate Gamma function}
\usage{
mvgamma(x, p = 1L)
}
\arguments{
\item{x}{The value to evaluate}

\item{p}{The dimension}
}
\description{
The multivariate gamma function.
}
