% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvn_E_lpdf}
\alias{mvn_E_lpdf}
\title{Calculate E_q[ln p(x)] where q(x) = MVN(x | mu, Sigma) and p(x) = MVN(x | mu0, Sigma0)}
\usage{
mvn_E_lpdf(mu0, Sigma0, mu, Sigma)
}
\arguments{
\item{mu0}{Mean prior}

\item{Sigma0}{Covariance prior}

\item{mu}{Variational mean}

\item{Sigma}{Variational covariance}
}
\description{
E_q[ln p(x)] where x ~ MVN(mu0, Sigma0) and q(x) = MVN(x | mu, Sigma)
}
