% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvn_entropy}
\alias{mvn_entropy}
\title{Return the entropy of multivariate normal density}
\usage{
mvn_entropy(S)
}
\arguments{
\item{S}{The covariate matrix}
}
\value{
value The entropy
}
\description{
Return the entropy of multivariate normal density
}
