% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ph_exponential}
\alias{ph_exponential}
\title{Perform Normal approximation variational inference for 
proportional-hazards model with exponential base.}
\usage{
ph_exponential(X, y, v, mu0, Sigma0, tol = 1e-08, maxiter = 100L)
}
\arguments{
\item{X}{The design matrix}

\item{y}{The response vector}

\item{v}{The censoring vector}

\item{mu0}{The prior mean for beta}

\item{Sigma0}{The prior covariance for beta}
}
\value{
v A list of relevant outputs
}
\description{
Perform Normal approximation variational inference for 
proportional-hazards model with exponential base.
}
