% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pnorm_mat}
\alias{pnorm_mat}
\title{Evaluate standard normal cdf for matrix of variates}
\usage{
pnorm_mat(m)
}
\arguments{
\item{m}{A matrix of variates}
}
\description{
Evaluate standard normal cdf for matrix of variates
}
