% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfvb_lm.R
\name{sample_vbdist}
\alias{sample_vbdist}
\title{Sample from MFVB fit}
\usage{
sample_vbdist(object, par = NULL, n_samples = 1000)
}
\arguments{
\item{object}{an object of class `mfvb`}

\item{par}{a character vector indicating which distributions to sample}
}
\value{
A matrix of draws from the indicated distributions
}
\description{
Sample from MFVB fit
}
