% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaled_inv_chisq_E}
\alias{scaled_inv_chisq_E}
\title{Scaled Inverse Chi squared E[x]}
\usage{
scaled_inv_chisq_E(nu, tau)
}
\arguments{
\item{nu}{shape}

\item{tau}{scale}
}
\description{
Calculate E[x] where x ~ Scaled-Inv-Chisq(nu, tau)
}
