% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaled_inv_chisq_E_inv}
\alias{scaled_inv_chisq_E_inv}
\title{Scaled Inverse Chi squared E[1/x]}
\usage{
scaled_inv_chisq_E_inv(nu, tau)
}
\arguments{
\item{nu}{shape}

\item{tau}{scale}
}
\description{
Calculate E[1/x] where x ~ Scaled-Inv-Chisq(nu, tau^2)
}
