% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaled_inv_chisq_E_log}
\alias{scaled_inv_chisq_E_log}
\title{Scaled Inverse Chi squared E[log(x)]}
\usage{
scaled_inv_chisq_E_log(nu, tau)
}
\arguments{
\item{nu}{shape}

\item{tau}{scale}
}
\description{
Calculate E[log(x)] where x ~ Scaled-Inv-Chisq(nu, tau).
}
