% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaled_inv_chisq_H}
\alias{scaled_inv_chisq_H}
\title{Scaled Inverse Chi squared H[x]}
\usage{
scaled_inv_chisq_H(nu, tau)
}
\arguments{
\item{nu}{shape}

\item{tau}{scale}
}
\description{
Calculate H[x] where x ~ Scaled-Inv-Chisq(nu, tau)
}
