% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_two_level_sparse}
\alias{solve_two_level_sparse}
\title{Solve two level sparse matrix problem.}
\usage{
solve_two_level_sparse(a1, A11, a2, A22, A12)
}
\arguments{
\item{x}{A list of matrices}
}
\description{
Solve two level sparse matrix problem.
}
