% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfvb_lm.R
\name{summary.mfvb}
\alias{summary.mfvb}
\title{Summarise MFVB fit}
\usage{
\method{summary}{mfvb}(object, ...)
}
\arguments{
\item{object}{an object of class `mfvb`}

\item{...}{Unused}
}
\description{
Summarise MFVB fit
}
