% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_vb_lin_reg}
\alias{update_vb_lin_reg}
\title{Update mean-field variational inference for 
a basic linear regression model.}
\usage{
update_vb_lin_reg(vb_fit, X, y, tol = 1e-08, maxiter = 100L, verbose = FALSE)
}
\arguments{
\item{vb_fit}{A previous fit}

\item{X}{The new design matrix}

\item{y}{The new response vector}

\item{tol}{Tolerance for convergence of the elbo}

\item{maxiter}{Maximum number of iterations allowed}

\item{verbose}{Print trace of the lower bound to console. Default is \code{FALSE}.}
}
\value{
v A list of relevant outputs
}
\description{
Update mean-field variational inference for 
a basic linear regression model.
}
