\name{varapproxr-package}
\alias{varapproxr-package}
\alias{varapproxr}
\docType{package}
\title{
\packageTitle{varapproxr}
}
\description{
\packageDescription{varapproxr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{varapproxr}
\packageIndices{varapproxr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{varapproxr}

Maintainer: \packageMaintainer{varapproxr}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
