% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vb_glmm}
\alias{vb_glmm}
\title{Variational Bayes for logistic mixed model.}
\usage{
vb_glmm(X, Z, y, mu_beta, sigma_beta, mu, sigma, Au = 1, Bu = 1,
  Bqu = 1, tol = 1e-08, maxiter = 100L, verbose = FALSE)
}
\arguments{
\item{X}{The design matrix}

\item{Z}{Group design matrix}

\item{y}{The response vector}

\item{mu_beta}{The prior mean for beta}

\item{sigma_beta}{The prior covariance for beta}

\item{mu}{Initial value for mu}

\item{sigma}{Initial value for sigma}

\item{Au}{The prior shape for u}

\item{Bu}{The prior scale for u}

\item{tol}{Tolerance level}

\item{maxiter}{Maximum iterations}

\item{verbose}{Print trace of the lower bound to console. Default is \code{FALSE}.}

\item{Aqu}{The initial value for Aqu}
}
\value{
A list containing:
\describe{
  \item{converged}{Indicator for algorithm convergence.}
  \item{elbo}{Vector of the ELBO sequence.} 
  \item{mu}{The optimised value of mu.}
  \item{sigma}{The optimised value of sigma.}
}
}
\description{
Variational Bayes for logistic mixed model.
}
