% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vb_lm}
\alias{vb_lm}
\title{Mean-field variational inference for a normal linear model.}
\usage{
vb_lm(
  X,
  y,
  mu0,
  Sigma0,
  a0 = 0.01,
  b0 = 0.01,
  prior = 1L,
  tol = 1e-08,
  maxiter = 100L,
  verbose = FALSE
)
}
\arguments{
\item{X}{The design matrix}

\item{y}{The response vector}

\item{mu0}{The prior mean for beta}

\item{Sigma0}{The prior covariance for beta}

\item{a0}{The scale hyper-parameter}

\item{b0}{The shape hyper-parameter}

\item{prior}{The prior to be used, `1` - inverse-gamma(a0,b0), `2`- half-t(a0,b0)}

\item{tol}{Tolerance for convergence of the elbo}

\item{maxiter}{Maximum number of iterations allowed}

\item{verbose}{Print trace of the lower bound to console. Default is \code{FALSE}.}
}
\value{
v A list of relevant outputs
}
\description{
Mean-field variational inference for a normal linear model.
}
