% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vb_lmm_randint}
\alias{vb_lmm_randint}
\title{Variational Bayes for linear mixed model.}
\usage{
vb_lmm_randint(
  X,
  Z,
  y,
  mu_beta,
  sigma_beta,
  mu,
  sigma,
  Aeps = 1,
  Beps = 1,
  Au = 1,
  Bu = 1,
  Bqeps = 1,
  Bqu = 1,
  tol = 1e-08,
  maxiter = 100L,
  verbose = FALSE,
  trace = FALSE
)
}
\arguments{
\item{X}{The design matrix}

\item{Z}{Group design matrix}

\item{y}{The response vector}

\item{mu_beta}{The prior mean for beta}

\item{sigma_beta}{The prior covariance for beta}

\item{mu}{Initial value for mu}

\item{sigma}{Initial value for sigma}

\item{Aeps}{The prior shape for sigma_eps}

\item{Beps}{The prior scale for sigma_eps}

\item{Au}{The prior shape for sigma_u}

\item{Bu}{The prior scale for sigma_u}

\item{Bqeps}{The intial value for Bqeps}

\item{Bqu}{The initial value for Bqu}

\item{tol}{Tolerance level}

\item{maxiter}{Maximum iterations}

\item{verbose}{Print trace of the lower bound to console. Default is \code{FALSE}.}

\item{trace}{Print a trace of `mu` to console.}
}
\value{
A list containing:
\describe{
  \item{converged}{Indicator for algorithm convergence.}
  \item{elbo}{Vector of the ELBO sequence.} 
  \item{mu}{The optimised value of mu.}
  \item{Sigma}{The optimised value of Sigma.}
}
}
\description{
Variational Bayes for linear mixed model.
}
\examples{
library(nlme)
X <- model.matrix( ~ age + factor(Sex, levels = c("Female", "Male")), data = Orthodont)
Z <- kronecker(diag(1, 27), rep(1, 4))
y <- Orthodont$distance
mu0 <- rep(0, ncol(X))
S0 <- diag(1e8, ncol(X))
mu <- rep(0, ncol(X) + ncol(Z))
S <- diag(1, ncol(X) + ncol(Z))
A <- 1/100
B <- 1/100
fit <- vb_lmm_randint(X, Z, y, mu0, S0, mu, S, A, B, A, B, verbose = TRUE)

}
