% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vb_lmm_randintslope}
\alias{vb_lmm_randintslope}
\title{Variational Bayes for linear mixed model (random intercept and coefficient only).}
\usage{
vb_lmm_randintslope(
  Xlist,
  Zlist,
  ylist,
  beta_mu0,
  beta_sigma0,
  nu_Omega0,
  lambda_Omega0,
  pr_Omega = 1L,
  sigma_a0 = 0.01,
  sigma_b0 = 0.01,
  pr_sigma = 1L,
  tol = 1e-08,
  maxiter = 500L,
  verbose = FALSE,
  trace = FALSE,
  streamlined = FALSE,
  use_elbo = TRUE
)
}
\arguments{
\item{Xlist}{A list of subject specific design matrices}

\item{Zlist}{A list of subject specific group matrices}

\item{ylist}{A list of subject specific responses}

\item{beta_mu0}{Prior mean for beta}

\item{beta_sigma0}{Prior covariance for beta}

\item{nu_Omega0}{Prior df for Omega}

\item{lambda_Omega0}{Prior scale matrix for Omega}

\item{pr_Omega}{Prior type for Omega - 1 is IW(nu, lambda), 2 is HW(nu, 2*nu*diag(1/lambda^2))}

\item{sigma_a0}{The first hyper-parameter for prior on sigma}

\item{sigma_b0}{The second hyper-parameter for prior on sigma}

\item{pr_sigma}{The prior to use for sigma_epsilon - 1 is IG(a0,b0) and 2 is Half-t(a0, b0)}

\item{tol}{Tolerance level for assessing convergence}

\item{maxiter}{Maximum number of fixed-update iterations}

\item{verbose}{Print trace of ELBO}

\item{trace}{Return trace of parameters beta, gamma}

\item{streamlined}{Use streamlined updates (more efficient if dim(Zlist) is large).}

\item{use_elbo}{Should the ELBO be calculated and used for convergence checks?}
}
\description{
Performs variational inference for random intercept and coefficient model.
Currently assumes that all groups have same number of parameters.
That is, that all Zlist elements are of equal dimension.
}
