% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vb_logistic_n}
\alias{vb_logistic_n}
\title{Variational inference for binomial logistic regression model}
\usage{
vb_logistic_n(X, y, n, mu0, Sigma0, mu_init, Sigma_init, tol = 1e-08,
  maxiter = 1000L, maxiter_jj = 25L, alg = "jj", verbose = FALSE)
}
\arguments{
\item{X}{The design matrix}

\item{y}{The response vector}

\item{n}{The trial vector}

\item{mu0}{The prior mean for beta paramter}

\item{Sigma0}{The prior variance for beta parameter}

\item{tol}{The tolerance level to assess convergence}

\item{maxiter}{The maximum number of iterations}

\item{maxiter_jj}{The maximum number of Jaakkola-Jordan 
iterations to initialise estimation}

\item{alg}{The algorithm used for final estimation 
of variational parameters. 
Must be one of \code{jj}, \code{sj}, or \code{kmw}.}
}
\value{
A list containing:
\describe{
  \item{\code{converged}}{Indicator for algorithm convergence.}
  \item{\code{jj_converged}}{Indicator for convergence of initial Jaakkola-Jordan iterations.}
  \item{\code{elbo}}{Vector of the ELBO sequence.} 
  \item{\code{mu}}{The optimised value of mu.}
  \item{\code{Sigma}}{The optimised value of Sigma.}
}
}
\description{
This is an experimental function to perform variational inference
for binomial logistic regression models.
}
\section{Details}{

  By default, the algorithm always intialises with Jaakkola-Jordan updates
  until convergence or \code{maxiter_jj}.
}

