% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vb_pois_reg}
\alias{vb_pois_reg}
\title{Perform mean-field variational inference for 
a Poisson regression model.}
\usage{
vb_pois_reg(X, y, n, mu0, Sigma0, tol = 1e-08, maxiter = 100L, verbose = FALSE)
}
\arguments{
\item{X}{The design matrix}

\item{y}{The response vector}

\item{n}{The offset term}

\item{mu0}{The prior mean for beta}

\item{Sigma0}{The prior covariance for beta}

\item{tol}{Tolerance for convergence of the elbo}

\item{maxiter}{Maximum number of iterations allowed}

\item{verbose}{Print trace of the lower bound to console. Default is \code{FALSE}.}
}
\value{
v A list of relevant outputs
}
\description{
Perform mean-field variational inference for 
a Poisson regression model.
}
