% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vcov.mfvb}
\alias{vcov.mfvb}
\title{Extract model coefficient posterior covariance matrix}
\usage{
\method{vcov}{mfvb}(object)
}
\arguments{
\item{object}{a `mfvb` model}
}
\description{
`vcov.mfvb` implements the `vcov` generic for objects of class `mfvb`.
}
