% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vmp_lm}
\alias{vmp_lm}
\title{Variational Message Passing for Normal linear model.}
\usage{
vmp_lm(n, X, y, mu0, Sigma0, A, maxiter = 100L, tol = 1e-10, verbose = TRUE)
}
\arguments{
\item{n}{Number of rows}

\item{X}{The design matrix n by d.}

\item{y}{The observation vector n by 1.}

\item{mu0}{The prior mean on coefficients}

\item{Sigma0}{The prior variance on coefficients}

\item{A}{The prior scale on variance}

\item{maxiter}{The maximum number of iterations}

\item{tol}{Tolerance for convergence}

\item{verbose}{Print trace of the ELBO}
}
\description{
Variational Message Passing for Normal linear model.
}
