% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{woodbury}
\alias{woodbury}
\title{Woodbury matrix identity}
\usage{
woodbury(A, B, C, D)
}
\arguments{
\item{A}{mat A}

\item{B}{mat B}

\item{C}{mat C}

\item{D}{mat D}
}
\description{
(A + BCD)^{-1}
}
