% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GaussianLikelihoodFragment}
\alias{GaussianLikelihoodFragment}
\title{Gaussian likelihood fragment update}
\usage{
GaussianLikelihoodFragment(n, XtX, Xty, yty, eta1_in, eta2_in)
}
\description{
Gaussian likelihood fragment update
}
