% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mfvb_lm}
\alias{mfvb_lm}
\title{mfvb_lm}
\usage{
mfvb_lm(
  formula,
  data,
  subset,
  weights,
  na.action,
  prior = NULL,
  contrasts = NULL,
  x = FALSE,
  y = FALSE,
  ...
)
}
\arguments{
\item{formula}{an object of class  [stats::formula].}

\item{data}{an optional data frame, list or environment 
(or object coercible by as.data.frame to a data frame) 
containing the variables in the model. If not found in data, 
the variables are taken from environment(formula), 
typically the environment from which lm is called.}

\item{prior}{a list with elements (mu0, Sigma0, a, b) giving the model prior parameters.}
}
\value{

}
\description{
Fit a linear regression model using mean-field variational Bayes approximation.
}
