% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvn_entropy}
\alias{mvn_entropy}
\title{Multivariate Normal Entropy}
\usage{
mvn_entropy(S)
}
\arguments{
\item{S}{Covariance matrix}
}
\description{
Calculate and return the entropy for multivariate distribution
with supplied covariance matrix.
}
