% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vb_lin_reg}
\alias{vb_lin_reg}
\title{Perform mean-field variational inference for 
basic linear regression model.}
\usage{
vb_lin_reg(X, y, mu0, Sigma0, a0, b0, tol = 1e-08, maxiter = 100L)
}
\arguments{
\item{X}{The design matrix}

\item{y}{The response vector}

\item{mu0}{The prior mean for beta}

\item{Sigma0}{The prior covariance for beta}
}
\value{
v A list of relevant outputs
}
\description{
Perform mean-field variational inference for 
basic linear regression model.
}
