% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vb_lmm}
\alias{vb_lmm}
\title{Variational Bayes for linear mixed model.}
\usage{
vb_lmm(
  y,
  X,
  Zlist,
  J,
  R,
  mu_beta0,
  Sigma_beta0,
  xi_sigma,
  Lambda_sigma,
  xi_k,
  Lambda_k,
  tol = 1e-08,
  maxiter = 100L,
  verbose = FALSE,
  trace = FALSE
)
}
\arguments{
\item{y}{The response vector}

\item{X}{The design matrix}

\item{Zlist}{Collection of group design matrices}

\item{J}{First dimension of each Z in Zlist (e.g. number of subjects/sites)}

\item{R}{Second dimension of each Z in Zlist (e.g. number of variables, intercept and slope would be R = 2)}

\item{mu_beta0}{The prior mean for beta}

\item{Sigma_beta0}{The prior covariance for beta}

\item{xi_sigma}{The first prior parameter for covariance Sigma}

\item{Lambda_sigma}{The second prior parameter for covariance Sigma}

\item{xi_k}{A vector of first covariance parameters for hierarchical covariance}

\item{Lambda_k}{A list of second covariance parameters for hierarchical covariance}

\item{tol}{Tolerance level}

\item{maxiter}{Maximum iterations}

\item{verbose}{Print trace of the lower bound to console. Default is \code{FALSE}.}

\item{trace}{Print a trace of `mu` to console.}
}
\value{
A list containing:
\describe{
  \item{converged}{Indicator for algorithm convergence.}
  \item{elbo}{Vector of the ELBO sequence.} 
  \item{mu}{The optimised value of mu.}
  \item{Sigma}{The optimised value of Sigma.}
}
}
\description{
Variational approximation for linear mixed model assuming Gaussian distribution priors on fixed and grouped effects.
Priors on the variance parameters are Scaled-Inverse-Wishart distributions.
}
